requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
Flask==3.1.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.1.3
